Split opaque type inference from generating obligations for said opaque types
